+++ src/main/java/tn/workbot/coco_marketplace/services/AgencyDeliveryManService.java
@@ -2,8 +2,13 @@ package tn.workbot.coco_marketplace.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import tn.workbot.coco_marketplace.entities.AgencyBranch;
 import tn.workbot.coco_marketplace.entities.AgencyDeliveryMan;
+import tn.workbot.coco_marketplace.entities.Request;
+import tn.workbot.coco_marketplace.entities.enmus.RequestStatus;
+import tn.workbot.coco_marketplace.repositories.AgencyBranchRepository;
 import tn.workbot.coco_marketplace.repositories.AgencyDeliveryManRepository;
+import tn.workbot.coco_marketplace.repositories.RequestRepository;
 import tn.workbot.coco_marketplace.services.interfaces.AgencyDeliveryManIService;
 
 import java.util.List;
@@ -13,6 +18,10 @@ import java.util.Optional;
 public class AgencyDeliveryManService implements AgencyDeliveryManIService {
     @Autowired
     AgencyDeliveryManRepository admr;
+    @Autowired
+    AgencyBranchRepository  abr;
+    @Autowired
+    RequestRepository rr;
 
     @Override
     public AgencyDeliveryMan addAgencyDeliveryMan(AgencyDeliveryMan agencyDeliveryMan) {
@@ -26,7 +35,22 @@ public class AgencyDeliveryManService implements AgencyDeliveryManIService {
 
     @Override
     public void removeAgencyDeliveryMan(Long id) {
-        admr.deleteById(id);
+
+        AgencyDeliveryMan man=admr.findById(id).get();
+        List<Request> requests= (List<Request>) rr.findAll();
+        for (Request r:requests) {
+            if(r.getAgencyDeliveryMan().getId().equals(man.getId()))
+            {
+                if(!r.getRequestStatus().equals(RequestStatus.APPROVED))
+                {
+                    if(rr.countApprovedDeliveryAgence(id)==0) {
+                        admr.deleteById(id);
+                    }
+                }
+
+            }
+
+        }
     }
 
     @Override
@@ -40,5 +64,13 @@ public class AgencyDeliveryManService implements AgencyDeliveryManIService {
     @Override
     public List<AgencyDeliveryMan> RetrieveAgencyDeliveryMen() {
         return (List<AgencyDeliveryMan>) admr.findAll();
+    }
+
+    @Override
+    public AgencyDeliveryMan AssignAgencyDeliveryManByBranch(AgencyDeliveryMan agencyDeliveryMan, Long Id) {
+        AgencyDeliveryMan agencyDeliveryMan1=admr.save(agencyDeliveryMan);
+        AgencyBranch agencyBranch=abr.findById(Id).get();
+        agencyDeliveryMan1.setAgencyBranch(agencyBranch);
+        return admr.save(agencyDeliveryMan1);
     }
 }
